refactor(docs): split docs-map generator tests by concern
